updated pitch and roll
