+++ conversions/conversions.py
@@ -1,7 +1,8 @@
 import pandas as pd
 import time
-from Filter.filter import Filter
+from filter.filter import Filter
 from conversions.conversion_factor_enum import Constants as constants
+from calculations.calculations import Calculations
 
 # Add dampening to factor into the wheel loads. Depends on the velocity
 # Dampening is proportional to velocity and it's a linear relationship
@@ -76,7 +77,6 @@ class Conversions:
             self.acel_data, "Z", 4, 30, 2)
         return self.acel_data
 
-
     def convert_acel_to_g(self):
         for i, row in self.acel_data.iterrows():
             self.acel_data.loc[i, "X"] = (row["X"]) * 0.53
@@ -90,3 +90,25 @@ class Conversions:
             linpot_data.loc[i, "Time"] = time.strftime(
                 "%H:%M:%S.{} %Z".format(mlsec), time.localtime(time_step)
             )
+
+    def generate_pitch_roll_df(self, low_pass_data) -> pd.DataFrame:
+
+        time = low_pass_data["Time"]
+        x = low_pass_data["X_lowpass"]  # Accelerometer x values
+        y = low_pass_data["Y_lowpass"]  # Accelerometer y values
+        z = low_pass_data["Z_lowpass"]  # Accelerometer z values
+
+        # Calculate pitch and roll angles
+        pitch = []
+        roll = []
+        for i in range(len(x)):
+            p, r = Calculations.calculate_pitch_roll_angles(x[i], y[i], z[i])
+            pitch.append(p)
+            roll.append(r)
+
+        pitch_roll_df = pd.DataFrame({
+            'Time': time,
+            'Pitch': pitch,
+            'Roll': roll
+        })
+        return pitch_roll_df
